requirements.txt
langchain==1.2.15
langchain-core==1.2.28
langchain-community==0.4.1
langchain-google-genai==4.2.1
langchain-huggingface==1.2.1
langchain-chroma==0.2.4
langchain-text-splitters==1.1.1
chromadb==0.6.3
sentence-transformers==3.0.1
pymupdf==1.24.10
streamlit==1.38.0
python-dotenv==1.0.1
ragas==0.2.6
datasets==2.21.0
pandas==2.2.2